requirements.txt
googletrans==4.0.0rc1
matplotlib==3.1.2
numpy==1.18.1
pandas==1.1.5
Pillow==9.3.0
PyYAML==6.0
sentence_transformers==2.2.2
torch==1.7.1+cu101
torchvision==0.8.2+cu101



